+++ server/src/controllers/CartController/createCartController.js
@@ -65,50 +65,4 @@ const createCart = async (req) => {
   }
 };
 
-
-const getCart = async () => {
-  try {
-
-    const carts = await Cart.findAll({
-      include: [
-        {
-          model: Product,
-          as: 'products',
-          attributes: ['id'],
-        },
-      ],
-    });
-    
-    return carts;
-
-  } catch (error) {
-    console.error('Error in getCart:', error);
-    throw error; // Error al handler
-  } 
-};
-
-const deleteCart = async (req) => {
-  try {
-    const { cartId } = req.params;
-
-    if (isNaN(cartId)) {
-      return { status: 400, data: { error: 'Invalid cart ID' } };
-    }
-
-    const cart = await Cart.findByPk(cartId);
-
-    if (!cart) {
-      return { status: 404, data: { error: 'Cart not found' } };
-    }
-
-    await cart.destroy();
-
-    return { status: 204 };
-  } catch (error) {
-    console.error('Error in deleteCart:', error);
-    return { status: 500, data: { error: 'Failed to delete cart' } };
-  }
-};
-
-
-module.exports = { createCart, getCart, deleteCart  };
+module.exports = { createCart };
